(Balan) StepTiming を 1byte から 2byte に変更 #47

+++ PRG/Firmware/Balan/Core/Src/Libraries/StepScheduler.cpp
@@ -76,7 +76,7 @@ int StepScheduler::RegisterBrick(uint8_t brickId, BrickType type)
 	return 0;
 }
 
-int StepScheduler::BeginPattern(uint32_t currentTick, uint8_t brickId, int patternId, uint8_t stepTiming, bool isRepeat)
+int StepScheduler::BeginPattern(uint32_t currentTick, uint8_t brickId, int patternId, uint16_t stepTiming, bool isRepeat)
 {
 	if (brickId >= BrickNum) {
 		return -1;
@@ -124,7 +124,7 @@ int StepScheduler::SetBrightness(uint8_t brickId, uint8_t brightness)
 	return 0;
 }
 
-int StepScheduler::SetStepTiming(uint8_t brickId, uint8_t stepTiming)
+int StepScheduler::SetStepTiming(uint8_t brickId, uint16_t stepTiming)
 {
 	if (brickId >= BrickNum) {
 		return -1;
