+++ PRG/Tools/Sprinkler/Sprinkler/PatternConstants.cs
@@ -13,7 +13,7 @@ public static class PatternConstants
         }
 
         // パターン終了までにかかる時間の取得(ミリ秒)
-        public int GetTime(byte stepTiming)
+        public int GetTime(ushort stepTiming)
         {
             return StepCount * stepTiming;
         }
